feat: support more english colour names
